Build and link OpenSSL, start work on sockets, windows fixes
DCO-1.1-Signed-off-by: Ellie <kittens@wobble.ninja>

+++ horse64/corelib/io.c
@@ -338,7 +338,7 @@ int iolib_open(
         );
     }
     int filedescr = _open_osfhandle(
-        (int32_t)f2, _O_RDONLY
+        (intptr_t)f2, _O_RDONLY
     );
     if (filedescr < 0) {
         CloseHandle(f2);
